add delete endpoint
