--- src/routes/+page.js
@@ -1,8 +1,0 @@
-
-    export async function load({fetch}){
-        const res = await fetch('/api')
-        const todos = await res.json()
-
-        return {todos}
-        
-    }
