--- src/routes/about/+page.svelte
@@ -1,1 +1,0 @@
-<h1>About</h1>
